+++ example/main.dart
@@ -4,11 +4,10 @@ void main() async {
   var client = Client();
   client.connect('localhost');
   var sub = client.sub('subject1');
-  await Future.delayed(Duration(seconds: 1));
-  client.pub('subject1', 'message1'.codeUnits);
+  client.pubString('subject1', 'message1');
   var msg = await sub.stream.first;
 
-  print(msg);
+  print(msg.payloadString);
   client.unSub(sub);
   client.close();
 }